Moves checkProblematicCharRanges to font_spec.js
